+++ src/compute-engine/compute-engine.ts
@@ -1702,12 +1702,7 @@ export class ComputeEngine implements IComputeEngine {
    * Calls `ce.function()`, `ce.number()` or `ce.symbol()` as appropriate.
    */
   box(
-    expr:
-      | NumericValue
-      | Decimal
-      | Complex
-      | [num: number, denom: number]
-      | SemiBoxedExpression,
+    expr: NumericValue | Decimal | Complex | SemiBoxedExpression,
     options?: { canonical?: CanonicalOptions; structural?: boolean }
   ): BoxedExpression {
     return box(this, expr, options);